1_graph_creation/canonical_references/requirements_loader.txt
google-cloud-spanner==3.40.1
